Update libtiff to 4.0.1.
Remove globals from tif_pixarlog.c

Re-apply commit 786269bafff68f1965b1536acef11f25e0cd44d0

Also reapply some portability changes to tiffiop.h.

No cluster differences.

+++ gs/tiff/test/strip.c
@@ -1,4 +1,4 @@
-/* $Id: strip.c,v 1.3.2.1 2009-01-01 17:52:51 bfriesen Exp $ */
+/* $Id: strip.c,v 1.4 2008/03/28 01:42:06 bfriesen Exp $ */
 
 /*
  * Copyright (c) 2004, Andrey Kiselev  <dron@ak4719.spb.edu>
@@ -171,11 +171,11 @@ failure:
 	TIFFClose(tif);
 openfailure:
 	fprintf (stderr, "Can't create test TIFF file %s:\n"
-"    ImageWidth=%u, ImageLength=%u, RowsPerStrip=%u, Compression=%d,\n"
+"    ImageWidth=%ld, ImageLength=%ld, RowsPerStrip=%ld, Compression=%d,\n"
 "    BitsPerSample=%d, SamplesPerPixel=%d, SampleFormat=%d,\n"
 "    PlanarConfiguration=%d, PhotometricInterpretation=%d.\n",
-		 name, width, length, rowsperstrip, compression,
-		 bps, spp, sampleformat, planarconfig,
+		 name, (long) width, (long) length, (long) rowsperstrip,
+                 compression, bps, spp, sampleformat, planarconfig,
 		 photometric);
 	return -1;
 }
@@ -249,11 +249,11 @@ failure:
 	TIFFClose(tif);
 openfailure:
 	fprintf (stderr, "Can't read test TIFF file %s:\n"
-"    ImageWidth=%u, ImageLength=%u, RowsPerStrip=%u, Compression=%d,\n"
+"    ImageWidth=%ld, ImageLength=%ld, RowsPerStrip=%ld, Compression=%d,\n"
 "    BitsPerSample=%d, SamplesPerPixel=%d, SampleFormat=%d,\n"
 "    PlanarConfiguration=%d, PhotometricInterpretation=%d.\n",
-		 name, width, length, rowsperstrip, compression,
-		 bps, spp, sampleformat, planarconfig,
+		 name, (long) width, (long) length, (long) rowsperstrip,
+                 compression, bps, spp, sampleformat, planarconfig,
 		 photometric);
 	return -1;
 }
@@ -263,8 +263,7 @@ write_scanlines(TIFF *tif, const tdata_t array, const tsize_t size)
 {
 	uint32		length, row;
 	tsize_t		scanlinesize, offset;
-
-	(void) size;
+        (void) size;
 
 	if (!TIFFGetField(tif, TIFFTAG_IMAGELENGTH, &length)) {
 		fprintf (stderr, "Can't get tag %d.\n", TIFFTAG_IMAGELENGTH);
@@ -280,7 +279,7 @@ write_scanlines(TIFF *tif, const tdata_t array, const tsize_t size)
 	for (offset = 0, row = 0; row < length; offset+=scanlinesize, row++) {
 		if (TIFFWriteScanline(tif, (char *)array + offset, row, 0) < 0) {
 			fprintf (stderr,
-				 "Can't write image data at row %u.\n", row);
+				 "Can't write image data at row %lu.\n", (long) row);
 			return -1;
 		}
         }
